+++ src/router/index.js
@@ -1,7 +1,7 @@
 import Vue from 'vue';
 import VueRouter from 'vue-router';
 import HomeView from '../views/HomeView.vue';
-import { isOnline } from '@/lib/pwa';
+import { isOnline, isDesktop, isUsingApp } from '@/lib/pwa';
 
 Vue.use(VueRouter);
 
@@ -31,6 +31,11 @@ const routes = [
     component: () => import('@/views/OfflineView.vue'),
   },
   {
+    path: '/install',
+    name: 'install',
+    component: () => import('@/views/InstallView.vue'),
+  },
+  {
     path: '/apod',
     name: 'apod',
     component: () => import('@/views/ApodView.vue'),
@@ -49,6 +54,11 @@ const router = new VueRouter({
 });
 
 router.beforeEach((to, from, next) => {
+  if (!isDesktop() && !isUsingApp() && to.name !== 'install') {
+    next({ name: 'install' });
+    return;
+  }
+
   if (!isOnline() && to.name !== 'offline') {
     next({ name: 'offline' });
     return;