+++ tests/unit/views/InstallView.spec.js
@@ -1,5 +1,5 @@
 import { shallowMount } from '@vue/test-utils';
-import HomeView from '@/views/HomeView.vue';
+import InstallView from '@/views/InstallView.vue';
 
 jest.mock('@/components/common/ErrorBoundary.vue', () => ({
   name: 'ErrorBoundary',
@@ -11,11 +11,11 @@ jest.mock('@/components/common/LandingHero.vue', () => ({
   template: '<div class="mock-landing-hero"></div>',
 }));
 
-describe('HomeView', () => {
+describe('InstallView', () => {
   let wrapper;
 
   beforeEach(() => {
-    wrapper = shallowMount(HomeView, {
+    wrapper = shallowMount(InstallView, {
       stubs: ['ErrorBoundary', 'LandingHero'],
     });
   });
@@ -31,8 +31,8 @@ describe('HomeView', () => {
     expect(wrapper.exists()).toBeTruthy();
   });
 
-  it('contains the "home-view" class', () => {
-    expect(wrapper.classes()).toContain('home-view');
+  it('contains the "install-view" class', () => {
+    expect(wrapper.classes()).toContain('install-view');
   });
 
   it('renders the mocked components', () => {
@@ -40,9 +40,10 @@ describe('HomeView', () => {
     expect(wrapper.find('landinghero-stub').exists()).toBeTruthy();
   });
 
-  it('does not render install button', () => {
+  it('passes down correct props to landing hero component', () => {
+    const landingHeroComponent = wrapper.find('landinghero-stub');
     expect(
-      wrapper.find('landinghero-stub').attributes('displayinstallbutton')
-    ).toBeFalsy();
+      landingHeroComponent.attributes('displayinstallbutton')
+    ).toBeTruthy();
   });
 });